+++ src/infra/http/controllers/user/edit-user-profile.controller.ts
@@ -6,10 +6,10 @@ import {
   Request,
   UseGuards,
 } from '@nestjs/common';
-import { Roles } from '@/core/shared/decorators/roles.decorator';
-import type { EditProfileDto } from '@/core/dto/user/edit-profile.dto';
-import type { AuthenticatedUserRequest } from '@/core/dto/user/get-user.dto';
-import type { ROLE } from '@/core/types';
+import { Roles } from '@/shared/decorators/roles.decorator';
+import type { EditProfileDto } from '@/shared/dto/user/edit-profile.dto';
+import type { AuthenticatedUserRequest } from '@/shared/dto/user/get-user.dto';
+import type { ROLE } from '@/shared/types';
 import { EditUserProfileService } from '@Services/user/edit-user-profile.service';
 import { JwtAuthGuard } from '@/infra/auth/auth.guard';
